Refactor out tests requiring caffe
Caffe is an unruly dependency, and it is troublesome to get working
both with Tensorflow (Since Tensorflow hogs the graphics card),
troublesome to install, and troublesome to get working in a python3
environment.  Hence, going to refactor out any and all caffe code
and potentially replace with fixtures where possible.