Make tests in vgg_test pass
Part of trying to ensure all my tests are both runnable and pass
correctly.  In vgg_test, the saving test used a no-longer-existing
set of Vgg19 weights.  Updated it to test using the frcnn weights
instead.  This test hasn't been run successfully in ages; it is now
up to date, and it found an error in networks/vgg.py, actually, that
has been fixed.  Also made vgg_test.py never print to stdout, but
instead print to stderr, so that it doesn't clutter test runs.